requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
elasticsearch==8.11.0
python-dotenv==1.0.0
pydantic==2.11.5
pydantic-settings==2.1.0
aiohttp==3.12.2
redis==6.2.0